test: create new test file for operations
